change access log discard count to metrics register mode. change metrics structure to avoid cycle import
